Refactor

+++ DiscordBot.py
@@ -3,6 +3,8 @@ import discord
 from discord.ext import commands, tasks
 
 MY_USER_ID = 206640727797530624
+SALES_CHANNEL_ID = 662929547427315722   # Channel ID to post the sales.
+ROLE_MENTION = '<@&662906994969280533>' # What role to mention.
 bot = commands.Bot(command_prefix='!')
 ozb = OzBargain()
 token = ''
@@ -29,17 +31,17 @@ async def delete(ctx, sale_id):
     if ctx.author.id != MY_USER_ID:
         return
 
-    if sale_id.isdigit():
-        if ozb.sale_exists(sale_id, ozb.load_sales()):
-            ozb.delete_sale(sale_id)
-            await ctx.send(f'Sale {sale_id} successfully deleted')
-            return
+    if sale_id.isdigit() and ozb.sale_exists(sale_id, ozb.load_sales()):
+        ozb.delete_sale(sale_id)
+        await ctx.send(f'Sale {sale_id} successfully deleted')
+        return
 
     await ctx.send('Invalid sale ID')
 
 
 @tasks.loop(minutes=1.0)
 async def send_sales():
+    # Get the sales from the database that haven't expired.
     db_sales = [x for x in ozb.load_sales() if not ozb.has_expired(x)]
     # Get the eBay sales from OzBargain that haven't expired.
     ozb_sales = [x for x in ozb.get_ebay_sales() if not ozb.has_expired(x)]
@@ -51,8 +53,8 @@ async def send_sales():
         ozb.save_sales(new_sales)
 
         # Send the sales, mentioning the role.
-        channel = bot.get_channel(662929547427315722)
-        await channel.send('<@&662906994969280533>')
+        channel = bot.get_channel(SALES_CHANNEL_ID)
+        await channel.send(ROLE_MENTION)
         for sale in new_sales:
             msg = f'{sale["link"].replace("goto", "node")}\n'
             await channel.send(msg)
